update QuakePakFile.cs and QuakePakFileSystem.cs namespace to reflect directory rename
